Standard Via cells are now generated when a CHD is created. Previously they were generated only when used while reading into the database.
In the prompt line text editor, pressing Bsp or clicking button 1 in
the prompt line now exits "firstinsert" (overwrite) mode.  The fact
that button 1 which resets cursor position didn't do this was a bug.
One could start editing a string, use btn 1 to move the cursor to the
center, use Bsp to erase chars to the beginning.  Typing a char would
erase the remaining line and add the new char only.