Revised instance numbering functions, new CDs flag maintains validity of instance sequence numbers.
Changed names of instance name functions.

In main state point selection of instances where 3 or more instances
would be selected, the filter pop-up no longer appears.  It was too
confusing and probably not that useful.  Only the selection pop-up
appears, allowing selection state of each instance to be set.

Esc pressed twice now deselects all in Xiv, as in Xic.  If the
selection pop-up is being shown, it will be updated for deselections.