WRspice:  new subdirectory ipc_demo contains the spclient demo application.  This used to be packaged as a separate tar file that a user could build without the WRspice source.  Now that the WRspice source is open, spclient is greatly simplified.  It builds with the same libraries as WRspice itself.
This needs work, however.  Need to support Apple, and native Windows,
and QT where X Windows may not be accessible.  The existing framework
assumes X Windows, as does wrspiced, which also needs work.

Presently, the user has to go to this directory and type "make".
Perhaps better to build spclient in bin with everything else.